Tools: Move bindings generator into `generator` folder.
