+++ cMake/FreeCadMacros.cmake
@@ -126,7 +126,7 @@ ENDMACRO(fc_target_copy_resource_flat)
 # It would be a bit cleaner to generate these files in ${CMAKE_CURRENT_BINARY_DIR}
 
 macro(generate_from_xml BASE_NAME)
-    set(TOOL_PATH "${CMAKE_SOURCE_DIR}/src/Tools/generate.py")
+    set(TOOL_PATH "${CMAKE_SOURCE_DIR}/src/Tools/bindings/generate.py")
     file(TO_NATIVE_PATH "${TOOL_PATH}" TOOL_NATIVE_PATH)
     file(TO_NATIVE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/${BASE_NAME}.xml" SOURCE_NATIVE_PATH)
 
@@ -147,7 +147,7 @@ macro(generate_from_xml BASE_NAME)
         COMMAND ${PYTHON_EXECUTABLE} "${TOOL_NATIVE_PATH}" --outputPath "${OUTPUT_NATIVE_PATH}" ${BASE_NAME}.xml
         MAIN_DEPENDENCY "${BASE_NAME}.xml"
         DEPENDS
-        "${CMAKE_SOURCE_DIR}/src/Tools/generateTemplates/templateClassPyExport.py"
+        "${CMAKE_SOURCE_DIR}/src/Tools/bindings/templates/templateClassPyExport.py"
         "${TOOL_PATH}"
         WORKING_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}"
         COMMENT "Building ${BASE_NAME}.h/.cpp out of ${BASE_NAME}.xml"
