+++ src/Tools/bindings/templates/templateCPPFile.py
@@ -3,13 +3,13 @@
 # (c) 2006 Juergen Riegel
 
 import template
-import generateBase.generateModel_Module
-import generateBase.generateTools
+import model.generateModel_Module
+import model.generateTools
 
 
 class TemplateCPPFile(template.ModelTemplate):
     def Generate(self):
-        generateBase.generateTools.ensureDir(self.outputDir)
+        model.generateTools.ensureDir(self.outputDir)
         print("Generate() App Dir")
