+++ src/Tools/bindings/templates/templateClassPyExport.py
@@ -5,8 +5,8 @@
 
 from . import template
 import os, sys
-import generateBase.generateModel_Module
-import generateBase.generateTools
+import model.generateModel_Module
+import model.generateTools
 
 
 class TemplateClassPyExport(template.ModelTemplate):
@@ -39,19 +39,19 @@ class TemplateClassPyExport(template.ModelTemplate):
             if not os.path.exists(inputDir + exportName + "Imp.cpp"):
                 file = open(outputImp, "wb")
                 print("TemplateClassPyExport", "TemplateImplement", file.name)
-                generateBase.generateTools.replace(self.TemplateImplement, locals(), file)
+                model.generateTools.replace(self.TemplateImplement, locals(), file)
                 file.close()
 
         outputCpp = outputDir + exportName + ".cpp"
         with open(outputCpp, "wb") as file:
             print("TemplateClassPyExport", "TemplateModule", file.name)
-            generateBase.generateTools.replace(self.TemplateModule, locals(), file)
+            model.generateTools.replace(self.TemplateModule, locals(), file)
 
         outputHeader = outputDir + exportName + ".h"
         with open(outputHeader, "wb") as file:
             print("TemplateClassPyExport", "TemplateHeader", file.name)
-            generateBase.generateTools.replace(self.TemplateHeader, locals(), file)
-            # file.write( generateBase.generateTools.replace(self.Template,locals()))
+            model.generateTools.replace(self.TemplateHeader, locals(), file)
+            # file.write( model.generateTools.replace(self.Template,locals()))
 
     TemplateHeader = """
 // This file is generated by src/Tools/generateTemplates/templateClassPyExport.py out of the XML file
