+++ src/Tools/bindings/templates/templateModule.py
@@ -7,7 +7,7 @@ from . import template, templateModuleApp
 
 class TemplateModule(template.ModelTemplate):
     def Generate(self):
-        print("generateBase.generateModel_Module.Generate()\n")
+        print("model.generateModel_Module.Generate()\n")
         App = templateModuleApp.TemplateModuleApp()
         App.outputDir = self.outputDir
         App.module = self.module
