+++ src/Tools/bindings/templates/templateModuleAppFeature.py
@@ -3,19 +3,19 @@
 # (c) 2006 Juergen Riegel
 
 from . import template
-import generateBase.generateModel_Module
-import generateBase.generateTools
+import model.generateModel_Module
+import model.generateTools
 
 
 class TemplateFeature(template.ModelTemplate):
     def Generate(self):
         file = open(self.outputDir + self.feature.Name + "Imp.cpp", "w")
-        generateBase.generateTools.replace(self.TemplateImplement, locals(), file)
+        model.generateTools.replace(self.TemplateImplement, locals(), file)
         file = open(self.outputDir + self.feature.Name + ".cpp", "w")
-        generateBase.generateTools.replace(self.TemplateModule, locals(), file)
+        model.generateTools.replace(self.TemplateModule, locals(), file)
         file = open(self.outputDir + self.feature.Name + ".h", "w")
-        generateBase.generateTools.replace(self.TemplateHeader, locals(), file)
-        # file.write( generateBase.generateTools.replace(self.Template,locals()))
+        model.generateTools.replace(self.TemplateHeader, locals(), file)
+        # file.write( model.generateTools.replace(self.Template,locals()))
 
     TemplateHeader = """
 #ifndef @self.module.Name.upper()@_FEATURE_@self.feature.Name.upper()@_H
