+++ src/Tools/bindings/templates/templateModuleAppMain.py
@@ -3,15 +3,15 @@
 # (c) 2006 Juergen Riegel
 
 from . import template
-import generateBase.generateModel_Module
-import generateBase.generateTools
+import model.generateModel_Module
+import model.generateTools
 
 
 class TemplateModuleAppMain(template.ModelTemplate):
     def Generate(self):
         file = open(self.outputDir + "/App" + self.module.Name + ".cpp", "w")
-        generateBase.generateTools.replace(self.Template, locals(), file)
-        # file.write( generateBase.generateTools.replace(self.Template,locals()))
+        model.generateTools.replace(self.Template, locals(), file)
+        # file.write( model.generateTools.replace(self.Template,locals()))
 
     Template = """
 /***************************************************************************
